#8: Refactored styles so that style types are the same as attribute types and can be parsed the same.  Also refactored lengths into core and added a length attribute type.  Added the "corner-radius" style on backgrounds and got rid of the button styles, since a radius was the only attribute on it.

+++ projects/MUIS/src/org/muis/core/MuisUtils.java
@@ -232,7 +232,7 @@ public class MuisUtils
 	 * @param transparency The transparency of the color to make
 	 * @return The resulting partially transparent color
 	 */
-	public static Color getColor(Color base, float transparency)
+	public static Color getColor(Color base, double transparency)
 	{
 		if(transparency == 0)
 			return base;
